major refactor of where components live, typo fixes
