+++ src/components/EmptyDay/emptyDayComponent.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './App.css';
+import '../App/App.css';
 
 export function EmptyDay(props) {
   return (
