+++ src/components/IconComponent/iconComponent.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 export function IconComponent() {
   return (
     <div>
-      <img src={require('./assets/apcIconBig.jpg')} alt="big icon" />
+      <img src={require('../../assets/apcIconBig.jpg')} alt="big icon" />
     </div>
   );
 }
